server/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.0
distro==1.9.0
eval_type_backport==0.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
groq==0.37.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mistralai==1.9.11
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
